# premier-league/demo/src/main/java/com/example/demo/model/Saison.java
package com.example.demo.model;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class Saison {
    private Integer id;
    private String nom;
    private boolean estTerminee;
}